+++ pypeit/images/maskimage.py
@@ -1,13 +1,12 @@
 """ Module for image mask related items """
-from collections import OrderedDict
 import numpy as np
 
 from pypeit.bitmask import BitMask
-from pypeit.core import procimg
-from pypeit import datamodel
+from collections import OrderedDict
 
 from IPython import embed
 
+# I am keeping this here to avoid circular imports as even core routines need occasional access
 
 class ImageBitMask(BitMask):
     """
@@ -37,204 +36,5 @@ class ImageBitMask(BitMask):
             ('EXTRACT', 'Pixel masked during local skysub and extraction')
         ])
         super(ImageBitMask, self).__init__(list(mask_dict.keys()), descr=list(mask_dict.values()))
-
-
-class ImageMask(datamodel.DataContainer):
-    """
-    Class to handle masks associated with an Image
-
-    Args:
-        bpm (np.ndarray):
-            Bad pixel mask (int)
-        crmask (np.ndarray, optional):
-            Cosmic Ray mask (boolean)
-
-    Attributes:
-        fullmask (np.ndarray):
-            The bitmask values for the full mask
-        pars (dict):
-            Used to hold parameters used when creating masks
-    """
-
-    bitmask = ImageBitMask()
-    version = '1.0.0'
-    datamodel = {
-        'bpm': dict(otype=np.ndarray, atype=np.integer, desc='Bad pixel mask'),
-        'crmask': dict(otype=np.ndarray, atype=np.bool_, desc='CR mask image'),
-        'fullmask': dict(otype=np.ndarray, atype=np.integer, desc='Full image mask'),
-    }
-
-    def __init__(self, bpm, crmask=None, fullmask=None):
-
-        # Setup the DataContainer
-        super(ImageMask, self).__init__({'bpm': bpm, 'crmask': crmask, 'fullmask': fullmask})
-
-    def _bundle(self):
-        """
-        Over-write default _bundle() method to restrict to fullmask only
-
-        Note:  crmask will not write to FITS as it is bool
-
-        Returns:
-            :obj:`list`: A list of dictionaries, each list element is
-            written to its own fits extension.
-        """
-        d = []
-        if self.fullmask is not None:
-            d.append(dict(fullmask=self.fullmask))
-        return d
-
-    def build_crmask(self, detector, par, image, rawvarframe, subtract_img=None):
-        """
-        Generate the CR mask frame
-
-        Mainly a wrapper to procimg.lacosmic
-
-        Args:
-            detector (:class:`pypeit.images.detector_container.DetectorContainer`):
-                Detector object
-            par (:class:`pypeit.par.pypeitpar.ProcessImagesPar`):
-                Parameters that dictate the processing of the images.  See
-                :class:`pypeit.par.pypeitpar.ProcessImagesPar` for the
-                defaults.
-            image (np.ndarray):
-                Image to identify CR's in
-            rawvarframe (np.ndarray):
-                Variance image
-            subtract_img (np.ndarray, optional):
-                If provided, subtract this from the image prior to CR detection
-
-        Returns:
-            np.ndarray: Copy of self.crmask (boolean)
-
-        """
-        use_img = image if subtract_img is None else image - subtract_img
-        # Run LA Cosmic to get the cosmic ray mask
-        # TODO -- In processrawimage we pass in nonlinear * saturation not saturation
-        self.crmask = procimg.lacosmic(use_img,
-                                  detector['saturation'],
-                                  detector['nonlinear'],
-                                  varframe=rawvarframe,
-                                  maxiter=par['lamaxiter'],
-                                  grow=par['grow'],
-                                  remove_compact_obj=par['rmcompact'],
-                                  sigclip=par['sigclip'],
-                                  sigfrac=par['sigfrac'],
-                                  objlim=par['objlim'])
-        # Return
-        return self.crmask.copy()
-
-    def build_mask(self, image, ivar, saturation=1e10,
-                   mincounts=-1e10, slitmask=None):
-        """
-        Return the bit value mask used during extraction.
-
-        The mask keys are defined by :class:`ScienceImageBitMask`.  Any
-        pixel with mask == 0 is valid, otherwise the pixel has been
-        masked.  To determine why a given pixel has been masked::
-
-            bitmask = ScienceImageBitMask()
-            reasons = bm.flagged_bits(mask[i,j])
-
-        To get all the pixel masked for a specific set of reasons::
-
-            indx = bm.flagged(mask, flag=['CR', 'SATURATION'])
-
-        Args:
-            image (np.ndarray):
-                Image
-            ivar (np.ndarray or None):
-                Inverse variance of the input image
-            saturation (float, optional):
-                Saturation limit in counts or ADU (needs to match the input image)
-            slitmask (np.ndarray, optional):
-                Slit mask image;  Pixels not in a slit are masked
-            mincounts (float, optional):
-
-        Returns:
-            numpy.ndarray: Copy of the bit value mask for the science image.
-        """
-        # Instatiate the mask
-        mask = np.zeros_like(image, dtype=self.bitmask.minimum_dtype(asuint=True))
-
-        # Bad pixel mask
-        if self.bpm is not None:
-            indx = self.bpm.astype(bool)
-            mask[indx] = self.bitmask.turn_on(mask[indx], 'BPM')
-
-        # Cosmic rays
-        if self.crmask is not None:
-            indx = self.crmask.astype(bool)
-            mask[indx] = self.bitmask.turn_on(mask[indx], 'CR')
-
-        # Saturated pixels
-        indx = image >= saturation
-        mask[indx] = self.bitmask.turn_on(mask[indx], 'SATURATION')
-
-        # Minimum counts
-        indx = image <= mincounts
-        mask[indx] = self.bitmask.turn_on(mask[indx], 'MINCOUNTS')
-
-        # Undefined counts
-        indx = np.invert(np.isfinite(image))
-        mask[indx] = self.bitmask.turn_on(mask[indx], 'IS_NAN')
-
-        if ivar is not None:
-            # Bad inverse variance values
-            indx = np.invert(ivar > 0.0)
-            mask[indx] = self.bitmask.turn_on(mask[indx], 'IVAR0')
-
-            # Undefined inverse variances
-            indx = np.invert(np.isfinite(ivar))
-            mask[indx] = self.bitmask.turn_on(mask[indx], 'IVAR_NAN')
-
-        if slitmask is not None:
-            indx = slitmask == -1
-            mask[indx] = self.bitmask.turn_on(mask[indx], 'OFFSLITS')
-
-        self.fullmask = mask
-        return self.fullmask.copy()
-
-    def update_mask_slitmask(self, slitmask):
-        """
-        Update a mask using the slitmask
-
-        Args:
-            slitmask (np.ndarray):
-                Slitmask with -1 values pixels *not* in a slit
-
-        """
-        # Pixels excluded from any slit.
-        indx = slitmask == -1
-        # Finish
-        self.fullmask[indx] = self.bitmask.turn_on(self.fullmask[indx], 'OFFSLITS')
-
-    def update_mask_cr(self, crmask_new):
-        """
-        Update the mask bits for cosmic rays
-
-        The original are turned off and the new
-        ones are turned on.
-
-        Args:
-            crmask_new (np.ndarray):
-                New CR mask
-        """
-        self.fullmask = self.bitmask.turn_off(self.fullmask, 'CR')
-        indx = crmask_new.astype(bool)
-        self.fullmask[indx] = self.bitmask.turn_on(self.fullmask[indx], 'CR')
-
-    def __repr__(self):
-        repr = '<{:s}: '.format(self.__class__.__name__)
-        # Image
-        rdict = {}
-        for attr in ['bpm', 'crmask', 'fullmask']:
-            if hasattr(self, attr) and getattr(self, attr) is not None:
-                rdict[attr] = True
-            else:
-                rdict[attr] = False
-        repr += ' images={}'.format(rdict)
-        repr = repr + '>'
-        return repr
