+++ pypeit/images/processrawimage.py
@@ -230,27 +230,21 @@ class ProcessRawImage(object):
         bpm = self.spectrograph.bpm(self.filename, self.det, shape=self.image.shape)
 
         # Extras
-        if 'extras' in process_steps:
-            self.build_rn2img()
-            self.build_ivar()
-            steps_copy.remove('extras')
+        self.build_rn2img()
+        self.build_ivar()
 
         # Generate a PypeItImage
         pypeitImage = pypeitimage.PypeItImage(self.image, ivar=self.ivar, rn2img=self.rn2img, bpm=bpm,
                                               detector=self.detector)
         # Mask(s)
         if 'crmask' in process_steps:
-            if 'extras' in process_steps:
-                var = utils.inverse(pypeitImage.ivar)
-            else:
-                var = np.ones_like(pypeitImage.image)
-            #
+            var = utils.inverse(pypeitImage.ivar)
             #pypeitImage.mask.build_crmask(self.spectrograph, self.det, self.par, pypeitImage.image, var)
             pypeitImage.mask.build_crmask(self.detector, self.par, pypeitImage.image, var)
             steps_copy.remove('crmask')
         nonlinear_counts = self.spectrograph.nonlinear_counts(self.detector,
                                                               apply_gain='apply_gain' in process_steps)
-        pypeitImage.mask.build_mask(pypeitImage.image, pypeitImage.ivar,
+        pypeitImage.build_mask(pypeitImage.image, pypeitImage.ivar,
                                saturation=nonlinear_counts, #self.spectrograph.detector[self.det-1]['saturation'],
                                mincounts=self.detector['mincounts'])
         # Error checking
