+++ pypeit/requirements.txt
@@ -13,6 +13,8 @@ PyYAML>=5.1
 configobj>=5.0.6
 scikit-learn>=0.20
 IPython>=7.2.0
+# How do we add python itself!?
+#python>=3.7
 #ginga>=2.9  -- Can't include ginga as pip will not grab the JXP fork
 
 # linetools
